Copy paste images (#217)
* feat(paste images): add images using `paste`

* feat(images): increase size when clicking on the image.

* fix(big image): remove `Inset`

* ui(tool content): fix multi model tools being a different colour

* wip(images): windows paste bug.

* Revert "ui(tool content): fix multi model tools being a different colour"

This reverts commit 9f361485b0c82e3b50f1267be054b930eae282c3.

* fix: broken revert.

* ui: add cursor to image.

* fix(images): set maximum to 5,
