+++ src/components/Dropzone/Dropzone.tsx
@@ -4,7 +4,6 @@ import { Cross1Icon, ImageIcon } from "@radix-ui/react-icons";
 import { DropzoneInputProps, FileRejection, useDropzone } from "react-dropzone";
 import { useAttachedImages } from "../../hooks/useAttachedImages";
 import { TruncateLeft } from "../Text";
-import { ImageFile } from "../../features/AttachedImages/imagesSlice";
 
 export const FileUploadContext = createContext<{
   open: () => void;
@@ -18,11 +17,11 @@ export const FileUploadContext = createContext<{
 export const DropzoneProvider: React.FC<
   React.PropsWithChildren<{ asChild?: boolean }>
 > = ({ asChild, ...props }) => {
-  const { insertImage, setError, setWarning } = useAttachedImages();
+  const { setError, processAndInsertImages } = useAttachedImages();
 
   const onDrop = useCallback(
     (acceptedFiles: File[], fileRejections: FileRejection[]): void => {
-      void processImages(acceptedFiles, insertImage, setError, setWarning);
+      processAndInsertImages(acceptedFiles);
 
       if (fileRejections.length) {
         const rejectedFileMessage = fileRejections.map((file) => {
@@ -34,7 +33,7 @@ export const DropzoneProvider: React.FC<
         setError(rejectedFileMessage.join("\n"));
       }
     },
-    [insertImage, setError, setWarning],
+    [processAndInsertImages, setError],
   );
 
   // TODO: disable when chat is busy
@@ -128,68 +127,3 @@ export const FileList = () => {
     </Flex>
   );
 };
-
-async function processImages(
-  files: File[],
-  onSuccess: (image: ImageFile) => void,
-  onError: (reason: string) => void,
-  onAbort: (reason: string) => void,
-) {
-  for (const file of files) {
-    try {
-      const scaledImage = await scaleImage(file, 800);
-      const fileForChat = {
-        name: file.name,
-        content: scaledImage,
-        type: file.type,
-      };
-
-      onSuccess(fileForChat);
-    } catch (error) {
-      if (error === "abort") {
-        onAbort(`file ${file.name} reading was aborted`);
-      } else {
-        onError(`file ${file.name} processing has failed`);
-      }
-    }
-  }
-}
-
-function scaleImage(file: File, maxSize: number): Promise<string> {
-  return new Promise((resolve, reject) => {
-    const reader = new FileReader();
-    reader.onload = () => {
-      const img = new Image();
-      img.onload = () => {
-        const canvas = document.createElement("canvas");
-        const ctx = canvas.getContext("2d");
-        if (ctx === null) {
-          reject(`canvas.getContext("2d"), returned null`);
-        }
-
-        let width = img.width;
-        let height = img.height;
-
-        if (width > height && width > maxSize) {
-          height = Math.round((height *= maxSize / width));
-          width = maxSize;
-        } else if (height >= width && height > maxSize) {
-          width = Math.round((width *= maxSize / height));
-          height = maxSize;
-        }
-
-        canvas.width = width;
-        canvas.height = height;
-        ctx?.drawImage(img, 0, 0, width, height);
-
-        resolve(canvas.toDataURL(file.type));
-      };
-      img.onerror = reject;
-      img.src = reader.result as string;
-    };
-
-    reader.onabort = () => reject("aborted");
-    reader.onerror = () => reject("error");
-    reader.readAsDataURL(file);
-  });
-}
